// src/pages/api/api-contracts/responses/User.ts
export type UserResponse  = {
    id: number;
    first_name: string;
    last_name: string;
    username: string;
    password: string;
    email: string;
    street_name: string;
    street_number: string;
    postal_code: string;
    country: string;
    city: string;
    role: string;
    created_at: string;
    updated_at: string | null;
}